Move apply to dask.utils
