+++ dask/array/linalg.py
@@ -6,9 +6,8 @@ import toolz
 
 from ..base import tokenize, wait
 from ..blockwise import blockwise
-from ..compatibility import apply
 from ..highlevelgraph import HighLevelGraph
-from ..utils import derived_from
+from ..utils import derived_from, apply
 from .core import dotmany, Array, concatenate
 from .creation import eye
 from .random import RandomState
